feat: re organize aotf codebase
