+++ apps/platform/src/components/AssociationsToolkit/components/DataUploader/DataUploader.jsx
@@ -36,9 +36,9 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { Link, Tooltip } from "ui";
 import * as XLSX from "xlsx";
 
-import useAotfContext from "../hooks/useAotfContext";
+import useAotfContext from "../../hooks/useAotfContext";
 import ValidationQuery from "./ValidationQuery.gql";
-import client from "../../../client";
+import client from "../../../../client";
 import NestedItem from "./NestedItem";
 
 const BorderAccordion = styled(Accordion)(({ theme }) => ({
@@ -245,7 +245,7 @@ const FileExample = ({ entity = "target", runAction }) => {
   );
 };
 
-function DataUploader({ fileStem }) {
+function DataUploader() {
   const [activeStep, setActiveStep] = useState(0);
   const [queryTermsResults, setQueryTermsResults] = useState(null);
   const { entityToGet, pinnedEntries, setPinnedEntries } = useAotfContext();
