+++ apps/platform/src/components/AssociationsToolkit/components/HeaderControls/RequiredControl.jsx
@@ -1,8 +1,8 @@
 import { useState, useEffect } from "react";
 import { Checkbox } from "@mui/material";
 import { styled } from "@mui/material/styles";
-import useAotfContext from "../hooks/useAotfContext";
-import { checkBoxPayload, getControlChecked } from "../utils";
+import useAotfContext from "../../hooks/useAotfContext";
+import { checkBoxPayload, getControlChecked } from "../../utils";
 
 const OTCheckbox = styled(Checkbox)`
   padding: 0;
