+++ apps/platform/src/components/AssociationsToolkit/components/Table/AggregationsRow.jsx
@@ -2,9 +2,9 @@ import { useState } from "react";
 import { styled, Grid } from "@mui/material";
 
 import AggregationsTooltip from "./AssocTooltip";
-import associationsColumns from "../static_datasets/dataSourcesAssoc";
-import prioritizationColumns from "../static_datasets/prioritisationColumns";
-import { groupViewColumnsBy } from "../utils";
+import associationsColumns from "../../static_datasets/dataSourcesAssoc";
+import prioritizationColumns from "../../static_datasets/prioritisationColumns";
+import { groupViewColumnsBy } from "../../utils";
 import { GridContainer } from "../layout";
 
 const AggregationsContainer = styled(GridContainer)({
