+++ apps/platform/src/components/AssociationsToolkit/components/Table/ColoredCell.jsx
@@ -1,4 +1,4 @@
-import { getScale } from "../utils";
+import { getScale } from "../../utils";
 import Tooltip from "./AssocTooltip";
 
 const getClassName = ({ globalScore, hasValue }) => {
