+++ apps/platform/src/components/AssociationsToolkit/components/TargetPrioritisationSwitch.jsx
@@ -1,8 +1,8 @@
 import ToggleButton from "@mui/material/ToggleButton";
 import ToggleButtonGroup from "@mui/material/ToggleButtonGroup";
 
-import { DISPLAY_MODE } from "./utils";
-import useAotfContext from "./hooks/useAotfContext";
+import { DISPLAY_MODE } from "../utils";
+import useAotfContext from "../hooks/useAotfContext";
 
 function TargetPrioritisationSwitch() {
   const { displayedTable, setDisplayedTable } = useAotfContext();
