+++ apps/platform/src/components/AssociationsToolkit/context/aotfActions.ts
@@ -1,4 +1,4 @@
-import { Action, ActionType, Pagination } from "./types";
+import { Action, ActionType, Pagination } from "../types";
 
 export function onPaginationChange(newPagination: Pagination): Action {
   return {
