--- apps/platform/src/components/AssociationsToolkit/index.js
@@ -1,13 +1,0 @@
-export { default as AdvanceOptionsMenu } from "./AdvanceOptionsMenu";
-export { default as TableAssociations } from "./Table/TableAssociations";
-export { default as TargetPrioritisationSwitch } from "./TargetPrioritisationSwitch";
-export { default as SearhInput } from "./SearchInput";
-export { default as DataDownloader } from "./DataDownloader";
-export { default as DataUploader } from "./DataUploader/DataUploader";
-export { default as AotfApiPlayground } from "./AotfApiPlayground";
-export {
-  default as AssociationsContext,
-  AssociationsStateProvider as AssociationsProvider,
-} from "./context/AssociationsStateContext";
-export { default as useAotfContext } from "./hooks/useAotfContext";
-export * from "./layout";
