+++ apps/platform/src/components/AssociationsToolkit/index.ts
@@ -1,0 +1,13 @@
+export { default as AdvanceOptionsMenu } from "./components/AdvanceOptionsMenu";
+export { default as TableAssociations } from "./components/Table/TableAssociations";
+export { default as TargetPrioritisationSwitch } from "./components/TargetPrioritisationSwitch";
+export { default as SearhInput } from "./components/SearchInput";
+export { default as DataDownloader } from "./components/DataDownloader";
+export { default as DataUploader } from "./components/DataUploader/DataUploader";
+export { default as AotfApiPlayground } from "./components/AotfApiPlayground";
+export {
+  default as AssociationsContext,
+  AssociationsStateProvider as AssociationsProvider,
+} from "./context/AssociationsStateContext";
+export { default as useAotfContext } from "./hooks/useAotfContext";
+export * from "./components/layout";
