+++ apps/platform/src/pages/DiseasePage/DiseaseAssociations/index.ts
